requirements.txt
torch
torchvision
flask
gradio
Pillow
scikit-image
numpy
natten 
einops